refactor: reorganize authentication components into auth directory and implement OAuthGitHub and OAuthGoogle
