+++ components/auth/sign-out-button.tsx
